Renamed 'perpetual' to 'auto' across the board
