+++ .project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>sz-sdk-java-perpetual</name>
+	<name>sz-sdk-java-auto</name>
 	<comment></comment>
 	<projects>
 	</projects>
