+++ src/test/java/com/senzing/sdk/core/auto/BasicRetryTest.java
@@ -1,4 +1,4 @@
-package com.senzing.sdk.core.perpetual;
+package com.senzing.sdk.core.auto;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
@@ -25,6 +25,7 @@ import com.senzing.sdk.SzRecordKey;
 import com.senzing.sdk.SzRecordKeys;
 import com.senzing.sdk.SzRetryableException;
 import com.senzing.sdk.core.SzCoreEnvironment;
+import com.senzing.sdk.core.auto.SzAutoCoreEnvironment;
 import com.senzing.sdk.test.StandardTestDataLoader;
 import com.senzing.sdk.test.SzRecord;
 import com.senzing.sdk.test.SzRecord.SzFullName;
@@ -84,14 +85,14 @@ import static com.senzing.util.JsonUtilities.*;
 @TestInstance(Lifecycle.PER_CLASS)
 @Execution(ExecutionMode.SAME_THREAD)
 @TestMethodOrder(OrderAnnotation.class)
-public class BasicRetryTest extends AbstractPerpetualCoreTest 
+public class BasicRetryTest extends AbstractAutoCoreTest 
 {
-    private static class MockEnvironment extends SzPerpetualCoreEnvironment {
+    private static class MockEnvironment extends SzAutoCoreEnvironment {
         private static ThreadLocal<List<SzException>> MOCK_FAILURES
             = new ThreadLocal<>();
 
         public MockEnvironment(String instanceName, String settings) {
-            super(SzPerpetualCoreEnvironment.newPerpetualBuilder()
+            super(SzAutoCoreEnvironment.newAutoBuilder()
                     .settings(settings).instanceName(instanceName)
                     .configRefreshPeriod(REACTIVE_CONFIG_REFRESH)
                     .concurrency(null));
