+++ src/test/java/com/senzing/sdk/core/auto/ConfigManagerTest.java
@@ -1,27 +1,31 @@
-package com.senzing.sdk.core.perpetual;
+package com.senzing.sdk.core.auto;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.parallel.Execution;
 import org.junit.jupiter.api.parallel.ExecutionMode;
 
+import com.senzing.sdk.SzException;
 import com.senzing.sdk.SzConfigManager;
 import com.senzing.sdk.SzEnvironment;
-import com.senzing.sdk.SzException;
 import com.senzing.sdk.test.StandardTestConfigurator;
-import com.senzing.sdk.test.SzConfigTest;
+import com.senzing.sdk.test.SzConfigManagerTest;
 import com.senzing.sdk.core.SzCoreEnvironment;
+import com.senzing.sdk.core.auto.SzAutoCoreEnvironment;
 
+import static org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import static org.junit.jupiter.api.TestInstance.Lifecycle;
 
 @TestInstance(Lifecycle.PER_CLASS)
 @Execution(ExecutionMode.SAME_THREAD)
-public class ConfigTest
-    extends AbstractPerpetualCoreTest 
-    implements SzConfigTest
+@TestMethodOrder(OrderAnnotation.class)
+public class ConfigManagerTest 
+    extends AbstractAutoCoreTest 
+    implements SzConfigManagerTest
 {
-    private SzPerpetualCoreEnvironment env = null;
+    private SzAutoCoreEnvironment env = null;
 
     private TestData testData = new TestData();
 
@@ -38,11 +42,13 @@ public class ConfigTest
     @BeforeAll
     public void initializeEnvironment() {
         this.beginTests();
-        this.initializeTestEnvironment();
+        this.initializeTestEnvironment(true);
+
         String settings = this.getRepoSettings();
         
         String instanceName = this.getClass().getSimpleName();
 
+        
         SzEnvironment env = SzCoreEnvironment.newBuilder()
                                              .instanceName(instanceName)
                                              .settings(settings)
@@ -54,12 +60,12 @@ public class ConfigTest
                 = new StandardTestConfigurator(env);
 
             this.testData.setup(configurator);
-            
+
         } finally {
             env.destroy();
         }
 
-        this.env = SzPerpetualCoreEnvironment.newPerpetualBuilder()
+        this.env = SzAutoCoreEnvironment.newAutoBuilder()
                                              .instanceName(instanceName)
                                              .settings(settings)
                                              .verboseLogging(false)
@@ -78,6 +84,6 @@ public class ConfigTest
             this.teardownTestEnvironment();
         } finally {
             this.endTests();
-        } 
+        }
     }
 }
