+++ src/test/java/com/senzing/sdk/core/auto/ConfigTest.java
@@ -1,31 +1,40 @@
-package com.senzing.sdk.core.perpetual;
+package com.senzing.sdk.core.auto;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.parallel.Execution;
 import org.junit.jupiter.api.parallel.ExecutionMode;
 
-import com.senzing.sdk.SzEngine;
+import com.senzing.sdk.SzConfigManager;
+import com.senzing.sdk.SzEnvironment;
 import com.senzing.sdk.SzException;
-import com.senzing.sdk.test.SzEngineBasicsTest;
+import com.senzing.sdk.test.StandardTestConfigurator;
+import com.senzing.sdk.test.SzConfigTest;
+import com.senzing.sdk.core.SzCoreEnvironment;
+import com.senzing.sdk.core.auto.SzAutoCoreEnvironment;
 
-import static org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import static org.junit.jupiter.api.TestInstance.Lifecycle;
 
-/**
- * Unit tests for {@link SzCoreDiagnostic}.
- */
 @TestInstance(Lifecycle.PER_CLASS)
 @Execution(ExecutionMode.SAME_THREAD)
-@TestMethodOrder(OrderAnnotation.class)
-public class EngineBasicsTest 
-    extends AbstractPerpetualCoreTest 
-    implements SzEngineBasicsTest
+public class ConfigTest
+    extends AbstractAutoCoreTest 
+    implements SzConfigTest
 {
+    private SzAutoCoreEnvironment env = null;
 
-    private SzPerpetualCoreEnvironment env = null;
+    private TestData testData = new TestData();
+
+    @Override
+    public SzConfigManager getConfigManager() throws SzException {
+        return this.env.getConfigManager();
+    }
+
+    @Override 
+    public TestData getTestData() {
+        return this.testData;
+    }
 
     @BeforeAll
     public void initializeEnvironment() {
@@ -34,8 +43,24 @@ public class EngineBasicsTest
         String settings = this.getRepoSettings();
         
         String instanceName = this.getClass().getSimpleName();
-        
-        this.env = SzPerpetualCoreEnvironment.newPerpetualBuilder()
+
+        SzEnvironment env = SzCoreEnvironment.newBuilder()
+                                             .instanceName(instanceName)
+                                             .settings(settings)
+                                             .verboseLogging(false)
+                                             .build();
+
+        try {
+            StandardTestConfigurator configurator
+                = new StandardTestConfigurator(env);
+
+            this.testData.setup(configurator);
+            
+        } finally {
+            env.destroy();
+        }
+
+        this.env = SzAutoCoreEnvironment.newAutoBuilder()
                                              .instanceName(instanceName)
                                              .settings(settings)
                                              .verboseLogging(false)
@@ -43,7 +68,7 @@ public class EngineBasicsTest
                                              .configRefreshPeriod(this.getConfigRefreshPeriod())
                                              .build();
     }
-    
+
     @AfterAll
     public void teardownEnvironment() {
         try {
@@ -54,16 +79,6 @@ public class EngineBasicsTest
             this.teardownTestEnvironment();
         } finally {
             this.endTests();
-        }
-    }
-
-    /**
-     * Gets the {@link SzEngine} from the {@link SzCoreEnvironment}.
-     * {@inheritDoc}
-     * 
-     * @return The {@link SzEngine} to use for this test.
-     */
-    public SzEngine getEngine() throws SzException {
-        return this.env.getEngine();
+        } 
     }
 }
