+++ src/test/java/com/senzing/sdk/core/auto/EngineBasicsTest.java
@@ -1,4 +1,4 @@
-package com.senzing.sdk.core.perpetual;
+package com.senzing.sdk.core.auto;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.AfterAll;
@@ -7,64 +7,36 @@ import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.parallel.Execution;
 import org.junit.jupiter.api.parallel.ExecutionMode;
 
+import com.senzing.sdk.SzEngine;
 import com.senzing.sdk.SzException;
-import com.senzing.sdk.SzConfigManager;
-import com.senzing.sdk.SzEnvironment;
-import com.senzing.sdk.test.StandardTestConfigurator;
-import com.senzing.sdk.test.SzConfigManagerTest;
-import com.senzing.sdk.core.SzCoreEnvironment;
+import com.senzing.sdk.core.auto.SzAutoCoreEnvironment;
+import com.senzing.sdk.test.SzEngineBasicsTest;
 
 import static org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import static org.junit.jupiter.api.TestInstance.Lifecycle;
 
+/**
+ * Unit tests for {@link SzCoreDiagnostic}.
+ */
 @TestInstance(Lifecycle.PER_CLASS)
 @Execution(ExecutionMode.SAME_THREAD)
 @TestMethodOrder(OrderAnnotation.class)
-public class ConfigManagerTest 
-    extends AbstractPerpetualCoreTest 
-    implements SzConfigManagerTest
+public class EngineBasicsTest 
+    extends AbstractAutoCoreTest 
+    implements SzEngineBasicsTest
 {
-    private SzPerpetualCoreEnvironment env = null;
 
-    private TestData testData = new TestData();
-
-    @Override
-    public SzConfigManager getConfigManager() throws SzException {
-        return this.env.getConfigManager();
-    }
-
-    @Override 
-    public TestData getTestData() {
-        return this.testData;
-    }
+    private SzAutoCoreEnvironment env = null;
 
     @BeforeAll
     public void initializeEnvironment() {
         this.beginTests();
-        this.initializeTestEnvironment(true);
-
+        this.initializeTestEnvironment();
         String settings = this.getRepoSettings();
         
         String instanceName = this.getClass().getSimpleName();
-
         
-        SzEnvironment env = SzCoreEnvironment.newBuilder()
-                                             .instanceName(instanceName)
-                                             .settings(settings)
-                                             .verboseLogging(false)
-                                             .build();
-
-        try {
-            StandardTestConfigurator configurator
-                = new StandardTestConfigurator(env);
-
-            this.testData.setup(configurator);
-
-        } finally {
-            env.destroy();
-        }
-
-        this.env = SzPerpetualCoreEnvironment.newPerpetualBuilder()
+        this.env = SzAutoCoreEnvironment.newAutoBuilder()
                                              .instanceName(instanceName)
                                              .settings(settings)
                                              .verboseLogging(false)
@@ -72,7 +44,7 @@ public class ConfigManagerTest
                                              .configRefreshPeriod(this.getConfigRefreshPeriod())
                                              .build();
     }
-
+    
     @AfterAll
     public void teardownEnvironment() {
         try {
@@ -84,5 +56,15 @@ public class ConfigManagerTest
         } finally {
             this.endTests();
         }
+    }
+
+    /**
+     * Gets the {@link SzEngine} from the {@link SzCoreEnvironment}.
+     * {@inheritDoc}
+     * 
+     * @return The {@link SzEngine} to use for this test.
+     */
+    public SzEngine getEngine() throws SzException {
+        return this.env.getEngine();
     }
 }
